[arcane] Déplace la définition de 'ComponentItem' et 'ComponentItemInternal' vers 'arcane/core/materials'.
